style: move AI Optimization settings from Providers tab to Languages tab
